+++ Sail-a-dex/Compatibility/PassageDude.cs
@@ -1,7 +1,6 @@
 ﻿using HarmonyLib;
-using System;
-using System.Linq;
-using System.Reflection;
+using static sailadex.SAD_Plugin;
+using static sailadex.Configs;
 
 namespace sailadex
 {
@@ -9,13 +8,10 @@ namespace sailadex
     {
         public static void PatchMod()
         {
-            Type ferryTravelClass = AppDomain.CurrentDomain.GetAssemblies()
-                .SelectMany(asm => asm.GetTypes())
-                .Where(type => type.IsClass && type.Name == "FerryTravel")
-                .Single();
-            MethodInfo original = AccessTools.Method(ferryTravelClass, "TeleportPlayer");
-            MethodInfo patch = AccessTools.Method(typeof(FerryTravelPatches), "TeleportPlayerPatch");
-            SAD_Plugin.harmony.Patch(original, new HarmonyMethod(patch));
+            var ferryTravelClass = AccessTools.TypeByName("FerryTravel");
+            var original = AccessTools.Method(ferryTravelClass, "TeleportPlayer");
+            var patch = AccessTools.Method(typeof(FerryTravelPatches), "TeleportPlayerPatch");
+            HarmonyInstance.Patch(original, new HarmonyMethod(patch));
         } 
 
         public class FerryTravelPatches
@@ -23,7 +19,7 @@ namespace sailadex
             [HarmonyPostfix]
             public static void TeleportPlayerPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value)
+                if (statsUIEnabled.Value)
                     StatsUI.Instance.PlayerTeleported();
             }
         }