+++ Sail-a-dex/Compatibility/RandomEncounters.cs
@@ -1,62 +1,66 @@
-﻿
-using BepInEx;
+﻿using BepInEx;
 using BepInEx.Configuration;
-using System;
-using System.Linq;
 using HarmonyLib;
 using System.Reflection;
+using static sailadex.SAD_Plugin;
 
 namespace sailadex
 {
     internal class RandomEncounters
     {
-        internal static BaseUnityPlugin pluginInstance;
-        internal static bool IsSeaLifeEnabled 
+        internal static bool IsSeaLifeModEnabled
         { 
             get 
             {
-                var seaLifeInstalled = pluginInstance.GetPrivateField<BaseUnityPlugin>("seaLifeModInstance");
-                var seaLifeControlled = pluginInstance.GetPrivateField<ConfigEntry<bool>>("controlSeaLifeMod");
-                return seaLifeInstalled != null && seaLifeControlled.Value;
+                var seaLifeInstalled = RE_PluginInstance.GetPrivateField<BaseUnityPlugin>("seaLifeModInstance");
+                var seaLifeControlled = RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("controlSeaLifeMod");
+                if (seaLifeControlled != null)
+                    return seaLifeInstalled != null && seaLifeControlled.Value;
+                return RE_PluginInstance.GetPrivateProperty<bool>("IsSeaLifeModEnabled");
             }
         }
 
-        internal static bool FlotsamEnabled { 
+        internal static bool IsFlotsamEnabled
+        { 
             get
             {
-                return pluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFlotsam").Value;
+                if (RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFlotsam") != null)
+                    return RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFlotsam").Value;
+                return RE_PluginInstance.GetPrivateProperty<bool>("IsFlotsamEnabled");
             }
         }
 
-        internal static bool DenseFogEnabled
+        internal static bool IsDenseFogEnabled
         {
             get 
             {
-                return pluginInstance.GetPrivateField<ConfigEntry<bool>>("enableDenseFog").Value;
+                if (RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableDenseFog") != null)
+                    return RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableDenseFog").Value;
+                return RE_PluginInstance.GetPrivateProperty<bool>("IsDenseFogEnabled");
             }
         }
-        internal static bool FishingBonanzaEnabled
+        internal static bool IsFishingBonanzaEnabled
         {
             get
             {
-                return pluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFishingBonanza").Value;
+                if (RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFishingBonanza") != null)
+                    return RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableFishingBonanza").Value;
+                return RE_PluginInstance.GetPrivateProperty<bool>("IsFishingBonanzaEnabled");
             }
         }
-        internal static bool IntenseStormEnabled
+        internal static bool IsIntenseStormEnabled
         {
             get
             {
-                return pluginInstance.GetPrivateField<ConfigEntry<bool>>("enableIntenseStorm").Value;
+                if (RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableIntenseStorm") != null)
+                    return RE_PluginInstance.GetPrivateField<ConfigEntry<bool>>("enableIntenseStorm").Value;
+                return RE_PluginInstance.GetPrivateProperty<bool>("IsIntenseStormEnabled");
             }
         }
 
         public static void PatchMod()
         {
-            Type encounterGeneratorClass = AppDomain.CurrentDomain.GetAssemblies()
-                .SelectMany(asm => asm.GetTypes())
-                .Where(type => type.IsClass && type.Name == "EncounterGenerator")
-                .Single();
-
+            var encounterGeneratorClass = AccessTools.TypeByName("EncounterGenerator");
             var methodsToPatch = new (string originalMethodName, string patchMethodName)[]
             {
                 ("GenerateFlotsam", "FlotsamCountPatch"),
@@ -65,12 +69,12 @@ namespace sailadex
                 ("GenerateFishingBonanza", "FishingBonanzaCountPatch"),
                 ("GenerateIntenseStorm", "IntenseStormCountPatch")
             };
-            
+
             foreach (var (originalMethodName, patchMethodName) in methodsToPatch)
             {
                 MethodInfo original = AccessTools.Method(encounterGeneratorClass, originalMethodName);
                 MethodInfo patch = AccessTools.Method(typeof(EncounterGeneratorPatches), patchMethodName);
-                SAD_Plugin.harmony.Patch(original, new HarmonyMethod(patch));
+                HarmonyInstance.Patch(original, new HarmonyMethod(patch));
             }
         }
 
@@ -79,35 +83,35 @@ namespace sailadex
             [HarmonyPostfix]
             public static void FlotsamCountPatch()
             {
-                if (FlotsamEnabled)
+                if (IsFlotsamEnabled)
                     StatsUI.Instance.IncrementIntStat("FlotsamEncounters");
             }
 
             [HarmonyPostfix]
             public static void DenseFogCountPatch()
             {
-                if (DenseFogEnabled)
+                if (IsDenseFogEnabled)
                     StatsUI.Instance.IncrementIntStat("DenseFogEncounters");
             }
 
             [HarmonyPostfix]
             public static void SeaLifeCountPatch()
             {
-                if (IsSeaLifeEnabled)
+                if (IsSeaLifeModEnabled)
                     StatsUI.Instance.IncrementIntStat("SeaLifeEncounters");
             }
 
             [HarmonyPostfix]
             public static void FishingBonanzaCountPatch()
             {
-                if (FishingBonanzaEnabled)
+                if (IsFishingBonanzaEnabled)
                     StatsUI.Instance.IncrementIntStat("FishingBonanzaEnc");
             }
 
             [HarmonyPostfix]
             public static void IntenseStormCountPatch()
             {
-                if (IntenseStormEnabled)
+                if (IsIntenseStormEnabled)
                     StatsUI.Instance.IncrementIntStat("IntenseStormEnc");
             }
         }