+++ Sail-a-dex/Patches/EventPatches.cs
@@ -1,22 +1,22 @@
-﻿using BepInEx.Logging;
-using HarmonyLib;
+﻿using HarmonyLib;
 using System.Linq;
 using UnityEngine;
+using static sailadex.SAD_Plugin;
+using static sailadex.Configs;
 
 namespace sailadex
 {
     internal class EventPatches
     {
-        static readonly ManualLogSource logger = SAD_Plugin.logger;
 
         [HarmonyPatch(typeof(FishingRodFish))]
         private class FishingRodFishPatches
         {
             [HarmonyPrefix]
             [HarmonyPatch("CollectFish")]
-            public static void CollectFishPatch(FishingRodFish __instance, GameObject ___currentFish)
+            public static void CollectFishPatch(GameObject ___currentFish)
             {
-                if (SAD_Plugin.fishCaughtUIEnabled.Value)
+                if (fishCaughtUIEnabled.Value)
                     FishCaughtUI.Instance.RegisterCatch(___currentFish);
             }
         }
@@ -27,14 +27,12 @@ namespace sailadex
 
             [HarmonyPostfix]
             [HarmonyPatch("OnTriggerEnter")]
-            public static void OnTriggerEnterPatch(IslandMarketWarehouseArea __instance, IslandMarket ___market, Collider other)
+            public static void OnTriggerEnterPatch(IslandMarket ___market, Collider other)
             {
-                if (SAD_Plugin.portsVisitedUIEnabled.Value && other.CompareTag("Player"))
+                if (portsVisitedUIEnabled.Value && other.CompareTag("Player"))
                     PortsVisitedUI.Instance.RegisterVisit(___market.GetPortName());
-                if (SAD_Plugin.statsUIEnabled.Value && other.CompareTag("Player"))
+                if (statsUIEnabled.Value && other.CompareTag("Player"))
                     StatsUI.Instance.IncrementPortVisited(___market.GetPortName());
-                
-                    
             }
         }
 
@@ -45,7 +43,7 @@ namespace sailadex
             [HarmonyPatch("EnterBoat")]
             public static void EnterBoatPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value && GameState.playing)
+                if (statsUIEnabled.Value && GameState.playing)
                     StatsUI.Instance.RegisterCurrentMass();
             }
 
@@ -53,7 +51,7 @@ namespace sailadex
             [HarmonyPatch("ExitBoat")]
             public static void ExitBoatPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value && GameState.playing)
+                if (statsUIEnabled.Value && GameState.playing)
                     StatsUI.Instance.RegisterCurrentMass();
             }
         }
@@ -76,12 +74,12 @@ namespace sailadex
             [HarmonyPatch("OnPickup")]
             public static void OnPickupPatch(Rigidbody ___boatRigidbody, string __state)
             {
-                if (SAD_Plugin.statsUIEnabled.Value &&
+                if (statsUIEnabled.Value &&
                     !GameState.currentlyLoading &&
                     GameState.playing &&
                     !___boatRigidbody.gameObject.GetComponent<BoatMooringRopes>().AnyRopeMoored())
                 {
-                    logger.LogDebug($"Unmoor from {__state} Day: {GameState.day} Time: {Sun.sun.globalTime}");
+                    LogDebug($"Unmoor from {__state} Day: {GameState.day} Time: {Sun.sun.globalTime}");
                     StatsUI.Instance.RegisterUnderway(__state);
                 }
             }
@@ -90,7 +88,7 @@ namespace sailadex
             [HarmonyPatch("MoorTo")]
             public static void MoorToPatch(Rigidbody ___boatRigidbody)
             {
-                if (SAD_Plugin.statsUIEnabled.Value &&
+                if (statsUIEnabled.Value &&
                     !GameState.currentlyLoading &&
                     GameState.playing &&
                     (___boatRigidbody.transform == GameState.lastBoat || ___boatRigidbody.transform == GameState.currentBoat?.parent) &&
@@ -100,7 +98,7 @@ namespace sailadex
                         .Where(r => r.GetPrivateField<SpringJoint>("mooredToSpring") != null)
                         .Select(r => r.GetPrivateField<SpringJoint>("mooredToSpring").transform.parent.name)
                         .FirstOrDefault();
-                    logger.LogDebug($"Moored at: {islandName} Day: {GameState.day} Time: {Sun.sun.globalTime} ");
+                    LogDebug($"Moored at: {islandName} Day: {GameState.day} Time: {Sun.sun.globalTime} ");
                     StatsUI.Instance.RegisterMoored(islandName);
                     StatsUI.Instance.UpdateMilesText();
                 }
@@ -114,7 +112,7 @@ namespace sailadex
             [HarmonyPatch("Drink")]
             public static void DrinkPatch(float ___health)
             {
-                if (SAD_Plugin.statsUIEnabled.Value && ___health > 0)
+                if (statsUIEnabled.Value && ___health > 0)
                     StatsUI.Instance.IncrementIntStat("DrinksTaken");
             }
         }
@@ -126,7 +124,7 @@ namespace sailadex
             [HarmonyPatch("EatFood")]
             public static void EatFoodPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value && !(PlayerNeeds.instance.eatCooldown > 0f))
+                if (statsUIEnabled.Value && !(PlayerNeeds.instance.eatCooldown > 0f))
                     StatsUI.Instance.IncrementIntStat("FoodsEaten");
             }
         }
@@ -138,7 +136,7 @@ namespace sailadex
             [HarmonyPatch("StopSmoking")]
             public static void StopSmokingPatch(float ___currentInhaleDuration)
             {
-                if (SAD_Plugin.statsUIEnabled.Value && ___currentInhaleDuration > 0f)
+                if (statsUIEnabled.Value && ___currentInhaleDuration > 0f)
                     StatsUI.Instance.IncrementIntStat("TimesSmoked");
             }
         }
@@ -150,11 +148,11 @@ namespace sailadex
             [HarmonyPatch("EnterBed")]
             public static void EnterBedPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value)
+                if (statsUIEnabled.Value)
                 {
                     StatsUI.Instance.IncrementIntStat("TimesSlept");
 
-                    if (SAD_Plugin.updateMilesSailed.Value == "sleep")
+                    if (updateMilesSailed.Value == "sleep")
                         StatsUI.Instance.UpdateMilesText();
                 }
             }
@@ -167,7 +165,7 @@ namespace sailadex
             [HarmonyPatch("CompleteMission")]
             public static void CompleteMissionPatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value)
+                if (statsUIEnabled.Value)
                     StatsUI.Instance.IncrementIntStat("MissionsCompleted");
             }
         }
@@ -179,7 +177,7 @@ namespace sailadex
             [HarmonyPatch("Update")]
             public static void UpdatePatch(bool ___teleportPlayer)
             {
-                if (SAD_Plugin.statsUIEnabled.Value && ___teleportPlayer)
+                if (statsUIEnabled.Value && ___teleportPlayer)
                     StatsUI.Instance.PlayerTeleported();
             }
         }
@@ -191,16 +189,15 @@ namespace sailadex
             [HarmonyPatch("FixedUpdate")]
             public static void FixedUpdatePatch()
             {
-                if (SAD_Plugin.statsUIEnabled.Value && GameState.currentBoat != null)                
+                if (statsUIEnabled.Value && GameState.currentBoat != null)
                     StatsUI.Instance.TrackDistance();
-                
             }
 
             [HarmonyPostfix]
             [HarmonyPatch("UpdateMass")]
             public static void UpdateMassPatch(Rigidbody ___body)
             {
-                if (SAD_Plugin.statsUIEnabled.Value && GameState.currentBoat != null)
+                if (statsUIEnabled.Value && GameState.currentBoat != null)
                     StatsUI.Instance.RegisterTotalMass(___body.mass);
             }
         }
@@ -212,7 +209,7 @@ namespace sailadex
             [HarmonyPatch("GetNormalizedDistance")]
             public static void GetNormalizedDistancePatch(float __result)
             {
-                if (SAD_Plugin.statsUIEnabled.Value &&
+                if (statsUIEnabled.Value &&
                     GameState.currentBoat != null && 
                     !GameState.currentBoat.parent.gameObject.GetComponent<BoatMooringRopes>().AnyRopeMoored() &&
                     __result <= WeatherStorms.instance.GetPrivateField<float>("rainBorder"))