+++ Sail-a-dex/Utilities/Extensions.cs
@@ -1,4 +1,5 @@
 ﻿using HarmonyLib;
+using System;
 
 namespace sailadex
 {
@@ -7,6 +8,11 @@ namespace sailadex
         public static T GetPrivateField<T>(this object obj, string field)
         {
             return (T)Traverse.Create(obj).Field(field).GetValue();
+        }
+
+        public static T GetPrivateProperty<T>(this object obj, string property)
+        {
+            return (T)Traverse.Create(obj).Property(property).GetValue();
         }
     }
 }